Updated vendor onboarding
